+++ benches/macrobenches.rs
@@ -59,7 +59,7 @@ fn parse_document_from_str(doc: &str) -> XmlResult<()> {
                 }
             }
             Event::Text(e) => {
-                black_box(e.xml_content()?);
+                black_box(e.xml10_content()?);
             }
             Event::CData(e) => {
                 black_box(e.into_inner());
@@ -84,7 +84,7 @@ fn parse_document_from_bytes(doc: &[u8]) -> XmlResult<()> {
                 }
             }
             Event::Text(e) => {
-                black_box(e.xml_content()?);
+                black_box(e.xml10_content()?);
             }
             Event::CData(e) => {
                 black_box(e.into_inner());
@@ -110,7 +110,7 @@ fn parse_document_from_str_with_namespaces(doc: &str) -> XmlResult<()> {
                 }
             }
             (resolved_ns, Event::Text(e)) => {
-                black_box(e.xml_content()?);
+                black_box(e.xml10_content()?);
                 black_box(resolved_ns);
             }
             (resolved_ns, Event::CData(e)) => {
@@ -138,7 +138,7 @@ fn parse_document_from_bytes_with_namespaces(doc: &[u8]) -> XmlResult<()> {
                 }
             }
             (resolved_ns, Event::Text(e)) => {
-                black_box(e.xml_content()?);
+                black_box(e.xml10_content()?);
                 black_box(resolved_ns);
             }
             (resolved_ns, Event::CData(e)) => {