+++ benches/microbenches.rs
@@ -150,7 +150,7 @@ fn one_event(c: &mut Criterion) {
             config.trim_text(true);
             config.check_end_names = false;
             match r.read_event() {
-                Ok(Event::Comment(e)) => nbtxt += e.xml_content().unwrap().len(),
+                Ok(Event::Comment(e)) => nbtxt += e.xml10_content().unwrap().len(),
                 something_else => panic!("Did not expect {:?}", something_else),
             };